:memo: Be a bit more verbose about the licensing
While the reuse compliance makes sure that the license can be determined automatically, the actual license was not very accessible.
Add a link to the license file and a hint to REUSE in the License section for easy and clear access to both.

Also state the copyright information here, as the license file is generic and it is easier to track the information here.